requirements.txt
numpy
scipy
imageio
opencv-contrib-python
pylint
pyserial

# for installation
setuptools
wheel
